과제 제출

+++ homework2-1.cpp
@@ -21,16 +21,16 @@ namespace IntClac
 
 namespace FloatClac
 {
-    float add(int a, int b) {
+    float add(float a, float b) {
         return a+b;
     }
-    float subtract(int a, int b) {
+    float subtract(float a, float b) {
         return a-b;
     }
-    float multiply(int a, int b) {
+    float multiply(float a, float b) {
         return a*b;
     }
-    float divide(int a, int b) {
+    float divide(float a, float b) {
         return a/b;
     }
 } // namespace FloatClac
